Don't wait for stake distribution
Local cluster starts for more than 3 shelley epochs, so stake
distribution is already working.
In tests we expect precise epochs numbers, so we need to get rid of
any unexpected (potential) waiting. If it happens in the future that we
have local cluster that doesn't have stake distribution working, let it
fail in an expected manner.