Move logfiles operation from cluster lock
Searching through log files is time consuming and blocks other operation
when running under global cluster lock.
Introduce a lock that is specific to logfiles search and rules files
cleanup.
A lock is needed so only one pytest worker is able to search logs at a
time and so the same error is not reported multiple times.