Wait for 2 new blocks on submit
Even though it should not be possible to submit the Tx, in reality the Tx is going through. Therefore we need to wait for two new blocks after submitting the Tx so we are sure that the input UTxO is really spent. Otherwise some other test may attempt to use the same UTxO and fail because of `BadInputsUTxO`.