Wait for log messages to get written to log files
The affected tests are very quick, and it can happen that the test is finished and log error ignore rule is delete before the actual error message gets written to a log file. As a result, it can happen that the error is not ignored and it is reported by the next test.