Fix race conditions in tests that use `add_ignore_rule`
When ignore rule is added after a command that produces the error has already run, another test can finish right after the error message is added to log file and before the ignore rule is created, and report the error.
Fix this race condition - the ignore rule has to be added before the error-producing command runs.