feat: add cleanup function and trap for SIGINT
This commit introduces a `_cleanup` function to stop all running cluster
instances and sets up a trap to call this function on SIGINT (Ctrl+C).
This ensures that the cluster instances are properly stopped even if the
script is interrupted. The `_cleanup` function is called at the end of
the script to stop the instances as well.