feat(tests): add conditional markers for Byron tests
Introduce ORDER5_BYRON and LONG_BYRON markers to conditionally apply
pytest marks based on the configuration. This change helps to avoid
setting priority for tests when the it is not needed, because the tests
run on setup where it is fast to respin the local testnet.

Also, add a new "noop" placeholder marker to the pyproject.toml file.