fix: handle improper coverage format with warning
The start-cluster.log can now contain also information other than
`cardano-cli` commands.

Previously, an AttributeError was raised when the coverage data was not in
the proper format. This change updates the code to log a warning message
instead and continue processing. This prevents the program from stopping
abruptly due to format issues.