refactor(cluster_scripts): store scripts_dir as instance attr
Store the result of get_testnet_variant_scriptdir2 as an instance attribute
(self.scripts_dir) in both LocalScripts and TestnetScripts classes. This
reduces redundant calls to the function and ensures consistent usage of the
scripts directory throughout the class methods. Updated all references to use
self.scripts_dir instead of re-calling the function.